% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_functions.R
\name{animate_pag}
\alias{animate_pag}
\alias{animate_pga}
\alias{animate_p}
\title{Animate the trajectory of devices}
\usage{
animate_pag(
  rst,
  dt,
  traj,
  cp,
  region = NULL,
  settings = mobvis_settings_animation(),
  title = NA,
  filename = "event_location_dev_\%s.mp4",
  width = 700,
  height = 700,
  fps = 3
)

animate_pga(
  rst,
  dt,
  traj,
  cp,
  region = NULL,
  settings = mobvis_settings_animation(),
  title = NA,
  filename = "posterior_dev_\%s.mp4",
  width = 700,
  height = 700,
  fps = 3
)

animate_p(
  rst,
  dt,
  traj,
  cp,
  region = NULL,
  settings = mobvis_settings_animation(),
  title = NA,
  filename = "prob_dev_\%s.mp4",
  width = 700,
  height = 700,
  fps = 3
)
}
\arguments{
\item{rst}{raster}

\item{dt}{data.table for \code{animate_pag} and \code{animate_pga} the following columns should be present: cell, rid, dist and either pag or pga. For \code{animate_p} the columns should be \code{t}, \code{dev}, \code{rid} and \code{p}.}

\item{traj}{sf object with the following columns: \code{t}, \code{cell}, \code{dev}, and \code{geometry} (the sf geometry column with points).}

\item{cp}{sf object of the cellplan}

\item{region}{sf object containing the region borders}

\item{settings}{settings of the animation. By default \code{\link{mobvis_settings_animation}}}

\item{title}{title of the animation. Can be either gif or mp4. Use the {percentage}s for the device id.}

\item{filename}{filename. Can be either gif or mp4. Use {percentage sign}s for the time id.}

\item{width}{width of the animation}

\item{height}{height of the animation}

\item{fps}{frame per second}
}
\description{
Animate the trajectory of devices, while showing the likelihood/event location probabilities (pag), the posterior probabilties (pg) or any other probabilities p. See the vignette of the simulator how they are used.
}
