% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_sig_strength}
\alias{map_sig_strength}
\alias{map_bast_server}
\alias{map_best_server}
\alias{map_pg}
\alias{map_pag}
\alias{map_p}
\alias{map_pga}
\title{Map signal strength (dBm), signal dominance (s), prior (pg), likelihood (pag), posterior (pga), and any other probability (p).}
\usage{
map_sig_strength(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  dev = NULL,
  type = c("dBm", "s"),
  interactive = TRUE,
  title = NA,
  settings = mobvis_settings(),
  ...
)

map_best_server(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  dev = NULL,
  interactive = TRUE,
  title = NA,
  settings = mobvis_settings(),
  ...
)

map_pg(
  rst,
  cp,
  region = NULL,
  dev = NULL,
  interactive = TRUE,
  title = NA,
  settings = mobvis_settings(),
  ...
)

map_p(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  dev = NULL,
  interactive = TRUE,
  title = NA,
  settings = mobvis_settings(),
  ...
)

map_pag(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  dev = NULL,
  interactive = TRUE,
  title = NA,
  settings = mobvis_settings(),
  ...
)

map_pga(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  dev = NULL,
  interactive = TRUE,
  title = NA,
  settings = mobvis_settings(),
  ...
)
}
\arguments{
\item{rst}{raster object}

\item{dt}{dt data.table that contains the data to be mapped}

\item{cp}{cellplan \code{sf} object of the cells}

\item{cells}{cells cells for which the data is shown}

\item{region}{region \code{sf} object that contains the borders of the region. If omitted, the raster borders are shown.}

\item{dev}{dev \code{sf} object of the location of one or more devices (needed for the animation functions)}

\item{type}{type \code{"dBm"} for signal strength or \code{"s"} for signal dominance}

\item{interactive}{should the map be interactive?}

\item{title}{title of the map. The default value is stored in the settings.}

\item{settings}{settings of the map. See \code{\link{mobvis_settings}}.}

\item{...}{arguments passed on to \code{\link[tmap:tm_raster]{tm_raster}}}
}
\description{
Map signal strength (dBm), signal dominance (s), prior (pg), likelihood (pag), posterior (pga) and any other probability (p). See the vignettes to see how there are used with the mobloc package and the simulator files.
}
