% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_mob_cells}
\alias{map_mob_cells}
\title{Low-level level to visualize mobile phone cells and auxiliary data}
\usage{
map_mob_cells(
  cp,
  rst,
  var = NULL,
  title = NA,
  palette = NA,
  cells = character(),
  region = NULL,
  dev = NULL,
  interactive = TRUE,
  opacity = 1,
  proxy = FALSE,
  settings = mobvis_settings(),
  ...
)
}
\arguments{
\item{cp}{cellplan}

\item{rst}{raster data}

\item{var}{if specified, \code{title} and \code{palette} are specified accordingly. Possible values: \code{"dBm", "s", "pga", "pag", "pg", "p", "bsm"}.}

\item{title}{title}

\item{palette}{palette The default depend is taken from the \code{settings}}

\item{cells}{cells to select}

\item{region}{borders (polygon) of the region of interest}

\item{dev}{\code{sf} object of device(s)}

\item{interactive}{should the map be interactive or static?}

\item{opacity}{the opacity of the raster layer.}

\item{proxy}{should the map be updated in a Shiny app?}

\item{settings}{mobvis settings}

\item{...}{arguments passed on to \code{\link[tmap:tm_raster]{tm_raster}}}
}
\description{
Visualize mobile phone cells and auxiliary data, such as signal strength, location estimation variables. This function is used in \code{\link{explore_mobloc}}, and also by dedicated functions such as \code{\link{map_sig_strength}}.
}
