% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{distance_plot}
\alias{distance_plot}
\alias{signal_dominance_plot}
\alias{radiation_plot}
\title{Signal strength model plots}
\usage{
distance_plot(W, ple, range, base_size = 11, show_classes = TRUE)

signal_dominance_plot(midpoint, steepness, base_size = 11, show_classes = TRUE)

radiation_plot(type = "a", beam_width, db_back = -30, base_size = 11)
}
\arguments{
\item{W}{power of an cell}

\item{ple}{path loss exponent}

\item{range}{range}

\item{base_size}{base size of the plot}

\item{show_classes}{show the class colors}

\item{midpoint}{middle point in the logistic function to map signal strength to probability}

\item{steepness}{width of the logistic function to map signal strength to probability}

\item{type}{\code{"a"} for azimuth (horizontal) plane and \code{"e"} for elevation/vertical plane}

\item{beam_width}{beam width}

\item{db_back}{difference in signal strength between the propagation direction of the cell and the opposite direction}
}
\description{
The \code{radiation_plot} plots the radiation in the horizontal (azimuth) or vertical (elevation) plane, the \code{distance_plot} the relation between distance and signal loss, and the \code{signal_dominance} plots the relation between signal strength and signal dominance, which is modelled as a logistic function. These plots are embedded in the interactive tool \code{\link{setup_sig_strength_model}}.
}
\examples{
########################
### distance_plot: plot signal strength over distance
########################

# Antenna of 5W, path loss exponent 2, which resembles free space
distance_plot(W = 5, ple = 2, range = 10000, show_classes = TRUE)

# Antenna of 5W, path loss exponent 4, which resembles urban area
distance_plot(W = 5, ple = 5, range = 1000, show_classes = TRUE)



########################
### signal_dominance: plot the relation between signal strength and the signal dominance,
### which is an indicator between 0 and 1 which represents the (modeled) value of connection
########################

# The default parameter values. The colors classes indicate the value of connection: from red
# (no/hardly connectivity, to yellow (fair connectivity) to blue/purple (excellent connectivity).
signal_dominance_plot(midpoint = -92.5, steepness = .2, show_classes = TRUE)

# explore the steepness parameter
gridExtra::grid.arrange(
    {signal_dominance_plot(-90, steepness = .1, show_classes = FALSE)},
    {signal_dominance_plot(-90, steepness = .25, show_classes = FALSE)},
    {signal_dominance_plot(-90, steepness = .5, show_classes = FALSE)},
    {signal_dominance_plot(-90, steepness = 1, show_classes = FALSE)},
    ncol = 1
)

########################
### radiation_plot: plot the relation between signal strength and the signal dominance,
### which is an indicator between 0 and 1 which represents the (modeled) value of connection
########################

# Azimuth plan pattern with a beam width of 64 degrees
# (so -32 and 32 degrees deviation means a signal loss of 3dB, the red dots)
radiation_plot(type = "a", beam_width = 64, db_back = -30)

# Elevation plan pattern with a beam width of 12 degrees
# (so -6 and 6 degrees deviation means a signal loss of 3dB, the red dots)
radiation_plot(type = "e", beam_width = 12, db_back = -30)
}
\seealso{
\code{\link{setup_sig_strength_model}}
}
