% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_sig_strength}
\alias{map_sig_strength}
\alias{map_bast_server}
\alias{map_best_server}
\alias{map_pg}
\alias{map_pag}
\alias{map_pga}
\title{Map signal strength (dBm), signal dominance (s), prior (pg), likelihood (pag), and posterior (pga).}
\usage{
map_sig_strength(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  type = c("dBm", "s"),
  interactive = TRUE,
  settings = mobvis_settings()
)

map_best_server(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  type = c("dBm", "s"),
  interactive = TRUE,
  settings = mobvis_settings()
)

map_pg(
  rst,
  cp,
  region = NULL,
  interactive = TRUE,
  settings = mobvis_settings()
)

map_pag(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  interactive = TRUE,
  settings = mobvis_settings()
)

map_pga(
  rst,
  dt,
  cp,
  cells = NA,
  region = NULL,
  interactive = TRUE,
  settings = mobvis_settings()
)
}
\arguments{
\item{rst}{rst}

\item{dt}{dt}

\item{cp}{cp}

\item{cells}{cells}

\item{region}{region}

\item{type}{type}

\item{interactive}{interactive}

\item{settings}{settings}
}
\description{
Map signal strength
}
