% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobvis_settings.R
\name{mobvis_settings}
\alias{mobvis_settings}
\alias{mobvis_settings_static}
\title{Settings of mobvis}
\usage{
mobvis_settings(
  titles = c(dBm = "Signal strength in dBm", s = "Signal dominance - s", bsm =
    "Best server map", pg = "Prior - P(g)", pag = "Connection likelihood - P(a|g)", pga =
    "Location posterior - P(g|a)"),
  palettes = list(dBm = "Blues", s = "-Blues", bsm = "Set2", pg = "-Blues", pag =
    "-Greens", pga = "viridis"),
  palette = "-Blues",
  cell_colors = c(Selected = "red", `Small cell` = "goldenrod3", `Normal cell` =
    "gold"),
  cell_size = 0.5,
  cell_offset = 150,
  cell_legend = length(cell_colors) > 1,
  cell_labels = FALSE,
  cell_label_color = "black",
  region.lwd = 2,
  use_classes = TRUE,
  dBm_classes = list(breaks = c(-Inf, seq(-120, -70, by = 10), Inf), labels =
    c("-120 or less", "-120 to -110", "-110 to -100", "-100 to -90", "-90 to -80",
    "-80 to -70", "-70 or better"), colors = RColorBrewer::brewer.pal(7, "Spectral"),
    lims = c(-120, -70), tit = "Signal strength (dBm)"),
  s_classes = list(breaks = seq(0, 1, by = 0.1), labels = paste(sprintf("\%.1f", seq(0,
    1, by = 0.1)[1:10]), "to", sprintf("\%.1f", seq(0, 1, by = 0.1)[2:11])), colors =
    RColorBrewer::brewer.pal(10, "Spectral"), lims = c(0, 1))
)

mobvis_settings_static(
  cell_colors = c(Selected = "red", `Small cell` = "black", `Normal cell` = "black"),
  cell_size = 0.5,
  cell_offset = 0,
  cell_legend = FALSE,
  region.lwd = 3,
  ...
)
}
\arguments{
\item{titles}{default titles. Named character vector, where the names should be dBm, s, bsm, pag, pg and pga.}

\item{palettes}{default palettes. Should be a list with the elements called: dBm, s, bsm, pag, pg and pga.}

\item{palette}{default palette for non-standard plots}

\item{cell_colors}{either a single color for all cells, or a vector of colors named: Selected, Small cell and Normal cell}

\item{cell_size}{size of the cell}

\item{cell_offset}{offset of the cells. If not 0, the cells are moved into the propagation direction}

\item{cell_legend}{should the legend for the different cells be included?}

\item{use_classes}{logical that determines whether the signal strength (dBm) and dominance (s) should be plot in classes (default) or using a sequential palette. The classes are defined by dBm_classes and s_classes}

\item{dBm_classes}{list that defines the classes for the signal strength}

\item{s_classes}{list that defines the classes for the signal dominance}

\item{...}{passed on to mobvis_settings}
}
\description{
Settings of mobvis. The default values of \code{mobvis_settings} are aimed for interactive maps, whereas the values for \code{mobvis_settings_static} are aimed for statis maps.
}
