% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{sim_get_region}
\alias{sim_get_region}
\alias{sim_get_raster}
\alias{sim_get_cellplan}
\alias{sim_get_signal_strength}
\title{Get simulator data}
\usage{
sim_get_region(sim)

sim_get_raster(sim)

sim_get_cellplan(sim)

sim_get_signal_strength(sim, rst, cp)
}
\arguments{
\item{sim}{list containing the following elements: \code{input_dir}, \code{output_dir}, \code{crs}.}

\item{rst}{rst}

\item{cp}{cp}
}
\description{
Get simulator data
}
